# openapi.yaml
openapi: 3.1.0
info:
  title: Vik Theatre API
  version: 0.1.0
  description: Student platform API. Single source of truth — generates Go server, TS client, Dart client.
servers:
  - url: https://api.viktheatre.in/v1
  - url: http://localhost:8080/v1

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Liveness probe
      security: []
      responses:
        "200": { description: ok }

  /auth/otp/send:
    post:
      summary: Send OTP to phone (MSG91)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpSendReq" }
      responses:
        "204": { description: sent }

  /auth/otp/verify:
    post:
      summary: Verify OTP, return JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpVerifyReq" }
      responses:
        "200":
          description: jwt pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }

  /me:
    get:
      summary: Current user
      responses:
        "200":
          description: user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }

  /students/{id}/channel:
    get:
      summary: Public/private student channel
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: channel
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Channel" }

  /admin/assets:
    post:
      summary: Create asset, returns Mux direct-upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetCreateReq" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetCreateRes" }

  /consent/{token}:
    get:
      summary: Fetch consent context (preview) for the parent consent UI
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: consent context
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset: { $ref: "#/components/schemas/Asset" }
                  student: { $ref: "#/components/schemas/User" }
                  batch_name: { type: string }
        "404": { description: invalid or expired token }
    post:
      summary: Parent signs consent
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConsentSignReq" }
      responses:
        "200": { description: signed }

  /admin/assets/{id}/publish:
    post:
      summary: Request publish / trigger consent flow for an asset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "202": { description: accepted — consent request dispatched }
        "404": { description: asset not found }

  /students/{id}/progress:
    get:
      summary: Student progress report — rubric averages + timeline
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: progress report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Progress" }

  /assets/{id}/rubric:
    get:
      summary: Rubric scores for a single asset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: rubric scores
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RubricScore" }

  /assets/{id}/notes:
    get:
      summary: List notes on an asset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: notes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Note" }
    post:
      summary: Add a note to an asset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                private: { type: boolean }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }

  /admin/assets/{id}/rubric:
    post:
      summary: Score an asset on one rubric dimension (admin/instructor)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RubricScore" }
      responses:
        "201": { description: scored }

  /admin/social/library:
    get:
      summary: List assets eligible for social posting (consented + public)
      responses:
        "200":
          description: library
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SocialLibraryItem" }

  /admin/social/posts:
    get:
      summary: List social posts for the current week
      responses:
        "200":
          description: posts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SocialPost" }
    post:
      summary: Create a social post (queued)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset_id, platforms, caption, scheduled_at]
              properties:
                asset_id: { type: string, format: uuid }
                platforms:
                  type: array
                  items: { $ref: "#/components/schemas/SocialPlatform" }
                caption: { type: string }
                scheduled_at: { type: string, format: date-time }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SocialPost" }

  /admin/social/posts/{id}/schedule:
    post:
      summary: Push a queued post to Buffer for scheduling
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: scheduled }

  /admin/assets/{id}/clips:
    get:
      summary: AI-suggested clips for a longer video asset
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: clip suggestions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClipSuggestion" }

  /payments/order:
    post:
      summary: Create a Razorpay order for a fee period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [period, amount_paise]
              properties:
                period: { type: string, example: "2026-05" }
                amount_paise: { type: integer, format: int64 }
      responses:
        "200":
          description: order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentOrder" }

  /payments:
    get:
      summary: List payments for the authenticated parent
      responses:
        "200":
          description: payments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Payment" }

  /payments/dues:
    get:
      summary: Pending due for the authenticated parent
      responses:
        "200":
          description: pending due
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dues" }

  /parent/consents:
    get:
      summary: List all consents for the logged-in parent
      responses:
        "200":
          description: consent list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConsentItem" }

  /parent/consents/pending:
    get:
      summary: Pending consent count for the logged-in parent
      responses:
        "200":
          description: pending count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }

  /admin/students:
    get:
      summary: List students for admin table
      responses:
        "200":
          description: student list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StudentListItem" }

  /notifications:
    get:
      summary: List current user's notifications (newest first, max 100)
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Notification' }

  /notifications/unread-count:
    get:
      summary: Get current user's unread notification count
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count: { type: integer }

  /notifications/{id}/read:
    post:
      summary: Mark a single notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: ok }

  /notifications/mark-all-read:
    post:
      summary: Mark all of the current user's notifications as read
      responses:
        "200": { description: ok }

  /webhooks/mux:
    post:
      summary: Mux webhook (asset.ready etc.)
      security: []
      responses: { "204": { description: ok } }

  /webhooks/razorpay:
    post:
      summary: Razorpay webhook
      security: []
      responses: { "204": { description: ok } }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    OtpSendReq:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: "+919876543210" }

    OtpVerifyReq:
      type: object
      required: [phone, code]
      properties:
        phone: { type: string }
        code:  { type: string, minLength: 4, maxLength: 8 }

    TokenPair:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:  { type: string }
        refresh_token: { type: string }

    User:
      type: object
      required: [id, role]
      properties:
        id:     { type: string, format: uuid }
        name:   { type: string }
        phone:  { type: string }
        email:  { type: string }
        role:   { type: string, enum: [student, parent, instructor, admin] }
        locale: { type: string, example: en }

    Channel:
      type: object
      properties:
        student:
          type: object
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            role: { type: string }
            locale: { type: string }
        assets:
          type: array
          items: { $ref: "#/components/schemas/Asset" }

    Asset:
      type: object
      properties:
        id:               { type: string, format: uuid }
        title:            { type: string }
        type:             { type: string, enum: [monologue, scene, showcase, catalog] }
        mux_playback_id:  { type: string }
        duration_s:       { type: integer }
        privacy:          { type: string, enum: [private, pending_consent, public] }
        created_at:       { type: string, format: date-time }

    AssetCreateReq:
      type: object
      required: [student_id, type, title]
      properties:
        student_id: { type: string, format: uuid }
        type:       { type: string, enum: [monologue, scene, showcase, catalog] }
        title:      { type: string }
        note:       { type: string }
        rubric:
          type: object
          additionalProperties: { type: integer, minimum: 0, maximum: 100 }

    AssetCreateRes:
      type: object
      properties:
        asset_id:           { type: string, format: uuid }
        mux_upload_url:     { type: string }

    RubricScore:
      type: object
      required: [dimension, score]
      properties:
        dimension: { type: string, example: voice_projection }
        score: { type: integer, minimum: 0, maximum: 100 }

    ConsentItem:
      type: object
      required: [id, asset_id, status]
      properties:
        id: { type: string, format: uuid }
        asset_id: { type: string, format: uuid }
        scope_channel: { type: boolean }
        scope_social: { type: boolean }
        scope_print: { type: boolean }
        scope_valid_months: { type: integer, enum: [6, 12, 24] }
        status: { type: string, enum: [pending, signed, revoked] }
        signed_at: { type: string, format: date-time }
        pdf_url: { type: string }

    StudentListItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string }
        batch_name: { type: string }
        asset_count: { type: integer }
        consent_status: { type: string, enum: [none, pending, signed] }

    ProgressDimension:
      type: object
      required: [dimension, score]
      properties:
        dimension: { type: string }
        score: { type: integer, minimum: 0, maximum: 100 }

    ProgressTimelineEntry:
      type: object
      required: [date, asset_id, asset_title, scores]
      properties:
        date: { type: string, format: date-time }
        asset_id: { type: string, format: uuid }
        asset_title: { type: string }
        scores:
          type: array
          items: { $ref: "#/components/schemas/ProgressDimension" }
        note: { type: string }

    Progress:
      type: object
      required: [averages, timeline]
      properties:
        averages:
          type: array
          items: { $ref: "#/components/schemas/ProgressDimension" }
        timeline:
          type: array
          items: { $ref: "#/components/schemas/ProgressTimelineEntry" }

    Note:
      type: object
      required: [id, asset_id, author_id, body, private, created_at]
      properties:
        id: { type: string, format: uuid }
        asset_id: { type: string, format: uuid }
        author_id: { type: string, format: uuid }
        body: { type: string }
        private: { type: boolean }
        created_at: { type: string, format: date-time }

    SocialPlatform:
      type: string
      enum: [ig_reel, yt_short, fb, linkedin]

    SocialLibraryItem:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        student_name: { type: string }
        mux_playback_id: { type: string }
        duration_s: { type: integer }
        thumbnail: { type: string }

    SocialPost:
      type: object
      required: [id, platforms, caption, status, created_at]
      properties:
        id: { type: string, format: uuid }
        asset_id: { type: string, format: uuid }
        platforms:
          type: array
          items: { $ref: "#/components/schemas/SocialPlatform" }
        caption: { type: string }
        scheduled_at: { type: string, format: date-time }
        buffer_id: { type: string }
        status: { type: string, enum: [queued, scheduled, posted, failed] }
        created_at: { type: string, format: date-time }
        asset_title: { type: string }
        student_name: { type: string }
        thumbnail: { type: string }

    ClipSuggestion:
      type: object
      required: [start_s, end_s, title, reason]
      properties:
        start_s: { type: integer }
        end_s: { type: integer }
        title: { type: string }
        reason: { type: string }

    Payment:
      type: object
      required: [id, user_id, amount_paise, status, created_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        razorpay_order_id: { type: string }
        amount_paise: { type: integer, format: int64 }
        status: { type: string, enum: [created, paid, failed] }
        period: { type: string, example: "2026-05" }
        created_at: { type: string, format: date-time }

    PaymentOrder:
      type: object
      required: [order_id, razorpay_key_id, amount_paise]
      properties:
        order_id: { type: string }
        razorpay_key_id: { type: string }
        amount_paise: { type: integer, format: int64 }

    Dues:
      type: object
      required: [period, amount_paise]
      properties:
        period: { type: string, example: "2026-05" }
        amount_paise: { type: integer, format: int64 }

    Notification:
      type: object
      required: [id, user_id, kind, title, meta, created_at]
      properties:
        id:         { type: string, format: uuid }
        user_id:    { type: string, format: uuid }
        kind:
          type: string
          enum: [consent_pending, asset_ready, feedback, class_reminder, fee_due]
        title:      { type: string }
        body:       { type: string, nullable: true }
        meta:       { type: object, additionalProperties: true }
        read_at:    { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }

    ConsentSignReq:
      type: object
      required: [scope]
      properties:
        otp:   { type: string }
        signed_name: { type: string }
        scope:
          type: object
          properties:
            channel: { type: boolean }
            social:  { type: boolean }
            print:   { type: boolean }
            valid_months: { type: integer, enum: [6, 12, 24] }